Move account-requests.js to aspsp-authorisation-server/account-requests.js
Move account-requests/account-request.js to aspsp-authorisation-server/account-request.js

Move test/acc-request-test.js to test/aspsp-authorisation-server/account-request-test.js

Configure --recursive flag to run tests in sub-directories.
