+++ lib/app.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const { OBAccountPaymentServiceProviders } = require('./ob-directory');
 const { authServer } = require('./aspsp-authorisation-server/token');
-const { accountRequests } = require('./account-requests');
+const { accountRequests } = require('./aspsp-authorisation-server/account-requests');
 const { openIdConfig } = require('./open-id-config');
 const bodyParser = require('body-parser');
