+++ lib/aspsp-authorisation-server/account-requests.js
@@ -1,6 +1,6 @@
 // See https://openbanking.atlassian.net/wiki/spaces/WOR/pages/5785171/Account+and+Transaction+API+Specification+-+v1.1.0#AccountandTransactionAPISpecification-v1.1.0-Request
 
-const { accountRequestHelper } = require('./account-requests/account-request.js');
+const { accountRequestHelper } = require('./account-request.js');
 
 const accountRequests = (() => {
   const post = (req, res) => {
